add new file for composable buttons also make logic for content type button
