+++ app/src/main/java/com/example/wepick/MainActivity.kt
@@ -69,67 +69,5 @@ fun StartProgram(navController: NavHostController, viewModel: MainViewModel) {
     )
 }
 
-@Composable
-fun NextButton(
-    navController: NavController,
-    modifier: Modifier = Modifier,
-    route: String,
-    enabled: Boolean,
-    onNextClick: () -> Unit
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    val isPressed by interactionSource.collectIsPressedAsState()
-
-    val shadowOffset = 4.dp
-    val targetOffset = if (isPressed) shadowOffset else 0.dp
-
-    val containerColor = if (enabled) AccentRed else Color(0xFFDDDDDD)
-    val contentColor = if (enabled) White else Color(0xFF111111).copy(alpha = 0.4f)
-
-    Box(
-        modifier = modifier
-            .padding(top = 16.dp)
-            .offset(x = targetOffset, y = targetOffset)
-            .drawBehind {
-                if (!isPressed && enabled) {
-                    drawRoundRect(
-                        color = Color(0xFF111111),
-                        topLeft = Offset(shadowOffset.toPx(), shadowOffset.toPx()),
-                        size = size,
-                        cornerRadius = CornerRadius(11.dp.toPx(), 11.dp.toPx())
-                    )
-                }
-            }
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                enabled = true,
-                onClick = {
-                    if (enabled) {
-                        onNextClick()
-                        navController.navigate(route)
-                    }
-                }
-            )
-            .background(
-                color = containerColor,
-                shape = RoundedCornerShape(11.dp)
-            )
-            .border(
-                width = 4.dp,
-                color = Color(0xFF111111),
-                shape = RoundedCornerShape(11.dp)
-            )
-            .padding(horizontal = 40.dp, vertical = 12.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            text = stringResource(R.string.next_button),
-            fontFamily = PressStart2P,
-            color = contentColor,
-            fontSize = 16.sp
-        )
-    }
-}
